TC-1500 - migrating the validation into the main checkin view
